Fix #7608
svn path=/trunk/koffice/; revision=59861

+++ kspread/kspread_interpreter.cc
@@ -1167,7 +1167,7 @@ static bool kspreadfunc_mod( KSContext& context )
 
   if ( !KSUtil::checkType( context, args[1], KSValue::DoubleType, true ) )
     return false;
-  if( args[1]->doubleValue()!=0)
+  if( (int)args[1]->doubleValue()!=0)
   {
         result=(int)args[0]->doubleValue() % (int)args[1]->doubleValue();
         context.setValue( new KSValue(  result  ) );
